Cloud build packages (#259)
* Create a deb-package-builder runtime builder.

Build debian packages using runtime builders.

1. copy from GCS bucket for php version
2. build php and extensions if necessary
3. copy output to GCS bucket

* Build the runtime builder.

* Add single bash script that can build all the extensions using cloud build.

* Skip downloading libuv-binaries if they already exist

* Fix the conditional for USE_LATEST

+++ deb-package-builder/extensions/cassandra-cpp-driver/build.sh
@@ -8,13 +8,17 @@ echo "Building cassandra cpp driver"
 
 PNAME="cassandra-cpp-driver"
 
-# Download the source
-download_from_tarball https://github.com/datastax/cpp-driver/archive/2.6.0.tar.gz 2.6.0
-PACKAGE_DIR=${PACKAGE_DIR}/packaging
+OUTPUT_FILE="${PNAME}_2.6.0-1~gcp8+1_amd64.deb"
 
-pushd ${PACKAGE_DIR}
-sed -i 's/libuv-dev/libuv1-dev/g' debian/control
-sed -i 's/release=1/release=1~gcp8+1/g' build_deb.sh
-./build_deb.sh
-cp build/*.deb ${ARTIFACT_DIR}
-popd
+if [ ! -f "${ARTIFACT_DIR}/${OUTPUT_FILE}" ]; then
+    # Download the source
+    download_from_tarball https://github.com/datastax/cpp-driver/archive/2.6.0.tar.gz 2.6.0
+    PACKAGE_DIR=${PACKAGE_DIR}/packaging
+
+    pushd ${PACKAGE_DIR}
+    sed -i 's/libuv-dev/libuv1-dev/g' debian/control
+    sed -i 's/release=1/release=1~gcp8+1/g' build_deb.sh
+    ./build_deb.sh
+    cp build/*.deb ${ARTIFACT_DIR}
+    popd
+fi
